refactor: remove metadata helpers, inline queries in tests
Drop helpers/metadata.ts and helpers/index.ts from both the template
and postcard example. Metadata queries are only used once in each test
file, so the extra abstraction layer wasn't pulling its weight.

Tests now import directly from helpers/client.ts and inline the
MetadataApiClient queries.

Made-with: Cursor
