--- packages/create-twenty-app/src/constants/template/src/__tests__/helpers/index.ts
@@ -1,2 +1,0 @@
-export { core, metadata, isCoreClientAvailable } from './client';
-export { findInstalledApp } from './metadata';
