--- packages/create-twenty-app/src/constants/template/src/__tests__/helpers/metadata.ts
@@ -1,17 +1,0 @@
-import { metadata } from './client';
-
-export async function findInstalledApp(universalIdentifier: string) {
-  const client = metadata();
-  const result = await client.query({
-    findManyApplications: {
-      id: true,
-      name: true,
-      universalIdentifier: true,
-    },
-  });
-
-  return result.findManyApplications.find(
-    (app: { universalIdentifier: string }) =>
-      app.universalIdentifier === universalIdentifier,
-  );
-}
