--- packages/twenty-apps/examples/postcard/src/__tests__/helpers/index.ts
@@ -1,3 +1,0 @@
-export { core, metadata, isCoreClientAvailable } from './client';
-export { findInstalledApp, findObjectByName } from './metadata';
-export { createPostCard, deletePostCard } from './mutations';
